requirements.txt
torch
pytorch-lightning
numpy
torchvision
opencv-python
Pillow
scikit-image
timm
